Lyric Offset handle float point numbers

+++ Models/Lyric.cs
@@ -10,7 +10,7 @@ public interface ILyric
     int StartTime { get; set; }
     string VideoId { get; set; }
     string Title { get; set; }
-    int Offset { get; set; }
+    float Offset { get; set; }
 }
 
 public class Lyric : ILyric
@@ -19,7 +19,7 @@ public class Lyric : ILyric
     public int StartTime { get; set; }
     public int LyricId { get; set; }
     public string Title { get; set; } = "";
-    public int Offset { get; set; }
+    public float Offset { get; set; }
 }
 
 class LyricConverter : JsonConverter<ILyric>
@@ -35,7 +35,7 @@ class LyricConverter : JsonConverter<ILyric>
             lyric.StartTime = queue.Dequeue().GetInt32();
             lyric.LyricId = queue.Dequeue().GetInt32();
             lyric.Title = queue.Dequeue().GetString() ?? "";
-            lyric.Offset = queue.Dequeue().GetInt32();
+            lyric.Offset = queue.Dequeue().GetSingle();
         }
         catch (InvalidOperationException) { }
